split bytecode.rs into multiple modules
